Family-agnostic ramps - refactor fee and family-specific logic to PriceRegistry (#1181)
## Motivation
Lifts out family-specific fee & validation logic to the PriceRegistry -
isolating fees calculations to the PriceRegistry state. The refactor
converts the OnRamp to be fully dest-chain family agnostic - future
additions of chain families would typically require only changing the
PriceRegistry.

## Solution
* Move the `DestChainConfig` to the PriceRegistry, and completely remove
dest chain configs from the OnRamp. Since most of the dynamic config is
specific to fee & chain logic (with the exception of `maxDataBytes` /
`maxNumberOfTokens`), it is optimal to store the configs in a single
contract
* Moves all fee token configs to the PriceRegistry
* Removes duplicate `_validateMessage`. Since the Router always calls
`getFee` before `forwardFromRouter` - we can do the validations early in
`getFee`
* `PriceRegistry.getValidatedFee` - performs the previous `getFee` logic
* `PriceRegistry.getValidatedRampMessageParams` - performs the
chain-specific logic of `forwardFromRouter` - validating pool return
data, computing message fees and determining out-of-order exec

![alt-chain-family-handling-design-Page-2 drawio
(1)](https://github.com/user-attachments/assets/79603dde-6531-4d17-bf8b-ae80eaf0dbfc)


### Contract Sizes

| Contract | Prev Size (kB) | Prev Margin (kB) | New Size (kB) | New
Margin (kB) |

|---------------------------------|------------------------------|---------------------------|---------------------------|---------------------------|
| EVM2EVMMultiOnRamp | 22.6 | 1.976 | 9.944 | 14.632 |
| PriceRegistry | 7.46 | 17.116 | 18.439 | 6.137 |
| **Total** | 30.06 | | 28.383 | |

### Gas costs (e2e)

| Function | Prev min-max | Prev avg | New min-max | New avg |

|--------------|--------------|---------------|----------------|---------|
| getFee (MultiOnRamp) | 0 - 47 501 | 20 357 | 0 - 38 863 | 16 655
| ccipSend (Router) | 287 347 - 375 060 | 341 643 | 279 728 - 367 441 |
334 024

---------

+++ contracts/src/v0.8/ccip/test/BaseTest.t.sol
@@ -25,7 +25,7 @@ contract BaseTest is Test {
   // Message info
   uint64 internal constant SOURCE_CHAIN_SELECTOR = 1;
   uint64 internal constant DEST_CHAIN_SELECTOR = 2;
-  uint64 internal constant GAS_LIMIT = 200_000;
+  uint32 internal constant GAS_LIMIT = 200_000;
 
   // Timing
   uint256 internal constant BLOCK_TIME = 1234567890;
